Add RBE support and migrate std::print to fmt::print
- Add build:rbe config for BuildBuddy Remote Build Execution
- Add toolchains_buildbuddy to MODULE.bazel (Ubuntu 24.04)
- Migrate std::print/println to fmt::print/println for GCC 13
- Fix std::powf to std::pow for standard compliance
- Update README to reflect GCC 13+ requirement

+++ src/lyra/mir/interp/runtime_real_ops.cpp
@@ -427,7 +427,7 @@ auto ShortRealLogicalNot(const RuntimeShortReal& op) -> RuntimeIntegral {
 
 auto ShortRealPower(const RuntimeShortReal& lhs, const RuntimeShortReal& rhs)
     -> RuntimeShortReal {
-  return {.value = std::powf(lhs.value, rhs.value)};
+  return {.value = std::pow(lhs.value, rhs.value)};
 }
 
 auto ShortRealToIntegral(
